Add Domain classes for Stock price monitoring
The design is to have a StockRecord class for a particular Stock.
StockRecord has the unique symbol for a partcular stock, and a
list of StockHistoryPriceRecord objects. StockHistoryPriceRecord
stores the price of a Stock and the time at which the price was
retrieved. A OneToMany relation connects StockRecord and
StockPriceHistoryRecord.